moving TurnController stuff into seperate file
