+++ src/game/CompactRules/CompactRules.ts
@@ -2,7 +2,8 @@ import { isPresent } from "ts-is-present";
 import { Direction, Gait, GaitParams, PieceName, PlayerName } from "../types";
 import { applyInSequence, keys } from "utilities";
 import { Board, Piece, Square } from "../Board";
-import { Game, TurnController } from "game/Game";
+import { Game } from "game/Game";
+import { TurnController } from "game/TurnController";
 import { Move, PieceDelta } from "game/Move";
 import { EventCenter } from "game/EventCenter";
 import { Pather, PatherParams } from "game/Pather";
